플레이어 이동속도 변환
- 네트워크캐릭터컨트롤러 MaxSpeed 값을 변환 시켰음

+++ Assets/3.Scripts/InGameA/Player/Player.cs
@@ -15,8 +15,8 @@ public class Player : NetworkBehaviour
     [System.Serializable]
     public class PlayerStats
     {
-        public int WalkSpeed;
-        public int RunSpeed;
+        public float WalkSpeed;
+        public float RunSpeed;
     }
     
     [Header("Character Stats")]
